+++ fusion-plugin-rpc-redux-react/src/__tests__/create-mock-emitter.js
@@ -1,0 +1,20 @@
+// @flow
+
+export default function createMockEmitter(props: mixed) {
+  const emitter = {
+    from: () => {
+      return emitter;
+    },
+    emit: () => {},
+    setFrequency: () => {},
+    teardown: () => {},
+    map: () => {},
+    on: () => {},
+    off: () => {},
+    mapEvent: () => {},
+    handleEvent: () => {},
+    flush: () => {},
+    ...props,
+  };
+  return emitter;
+}
